feat: refactor F0FileItem and F0NotesTextEditor components
