+++ packages/react/src/components/RichText/F0RichTextEditor/F0RichTextEditor.tsx
@@ -21,6 +21,7 @@ import {
 } from "@/components/RichText/internal"
 import { F0Button } from "@/components/F0Button"
 import { Cross } from "@/icons/app"
+import { experimentalComponent } from "@/lib/experimental"
 import { useI18n } from "@/lib/providers/i18n/i18n-provider"
 import { withSkeleton } from "@/lib/skeleton"
 import { cn } from "@/lib/utils"
@@ -47,11 +48,69 @@ import {
   handleEditorUpdate,
   setupContainerObservers,
 } from "./utils/helpers"
+import type { MentionsConfig } from "@/components/RichText/internal"
+import type { TranscribeFn } from "@/sds/ai/F0AiChat/types"
 import type {
-  F0RichTextEditorHandle,
-  F0RichTextEditorProps,
-  F0RichTextEditorSkeletonProps,
-} from "./types"
+  enhanceConfig,
+  filesConfig,
+  heightType,
+  primaryActionType,
+  resultType,
+  secondaryActionsType,
+} from "./utils/types"
+
+// Declared next to the component (not in a sibling types.ts) so api-extractor
+// rolls them into the bundled d.ts instead of emitting a broken relative import.
+export interface F0RichTextEditorProps {
+  mentionsConfig?: MentionsConfig
+  enhanceConfig?: enhanceConfig
+  filesConfig?: filesConfig
+  secondaryAction?: secondaryActionsType
+  primaryAction?: primaryActionType
+  onChange: (result: resultType) => void
+  maxCharacters?: number
+  placeholder: string
+  initialEditorState?: {
+    content?: string
+    files?: File[]
+  }
+  title: string
+  height?: heightType
+  plainHtmlMode?: boolean
+  fullScreenMode?: boolean
+  onFullscreenChange?: (fullscreen: boolean) => void
+  /** Whether the editor is disabled */
+  disabled?: boolean
+  /** Whether the editor has an error state */
+  error?: boolean
+  /** Whether the editor is in a loading state */
+  loading?: boolean
+  /**
+   * Voice dictation: transcribes a recorded audio blob into text inserted at
+   * the cursor. Same contract as F0AiChatTextArea — when omitted, the
+   * microphone button is not rendered.
+   */
+  onTranscribe?: TranscribeFn
+}
+
+export type F0RichTextEditorHandle = {
+  clear: () => void
+  clearFiles: () => void
+  focus: () => void
+  setError: (error: string | null) => void
+  setContent: (content: string) => void
+}
+
+export interface F0RichTextEditorSkeletonProps {
+  rows?: number
+}
+
+/** @deprecated Use F0RichTextEditorProps */
+export type RichTextEditorProps = F0RichTextEditorProps
+/** @deprecated Use F0RichTextEditorHandle */
+export type RichTextEditorHandle = F0RichTextEditorHandle
+/** @deprecated Use F0RichTextEditorSkeletonProps */
+export type RichTextEditorSkeletonProps = F0RichTextEditorSkeletonProps
 
 /** How long dictation errors stay visible (mirrors F0AiChatTextArea) */
 const DICTATION_ERROR_MS = 4000
@@ -533,7 +592,20 @@ const F0RichTextEditorSkeleton = ({
   )
 }
 
-export const F0RichTextEditor = withSkeleton(
-  F0RichTextEditorComponent,
-  F0RichTextEditorSkeleton
+/**
+ * @experimental This is an experimental component, use it at your own risk
+ */
+export const F0RichTextEditor = experimentalComponent(
+  "F0RichTextEditor",
+  withSkeleton(F0RichTextEditorComponent, F0RichTextEditorSkeleton)
 )
+
+/** @deprecated Use F0RichTextEditor */
+export const RichTextEditor = F0RichTextEditor
+
+export * from "./utils/constants"
+export * from "./utils/types"
+export type {
+  MentionedUser,
+  MentionsConfig,
+} from "@/components/RichText/internal/Extensions/Mention/types"
